+++ App/Screens/Income.js
@@ -1,26 +1,56 @@
 import React, { useContext, useState } from 'react';
-import { Alert, Keyboard, StatusBar, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
+import { Alert, Button, FlatList, Keyboard, KeyboardAvoidingView, StatusBar, StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
 import { AppContext } from './../Context/AppContext';
+import IncomeCatagories from './IncomeCatagories';
 import ModalForUser from './modal';
 
 const Income = () => {
     const { modal, setModal } = useContext(AppContext)
-
     const [newCatagory, setNewCatagory] = useState("")
-    const [subCatagories, setSubCatagories] = useState([])
+    const [subCatagories, setSubCatagories] = useState("")
+
+    const [addCatagoryModalactive, setAddCatagoryModalactive] = useState(false)
+    const [addTrasactionModalActive, setAddTrasactionModalActive] = useState(false)
 
     const addSubCatagory = () => {
         if (newCatagory == "") {
             Alert.alert("Blank Input", "Please write something")
         } else {
-            const newSubCatagory = {
-                id: Math.random(),
-                name: newCatagory
-            }
             Keyboard.dismiss();
-            setSubCatagories([...subCatagories, newSubCatagory])
+            setSubCatagories([...subCatagories, newCatagory])
             setNewCatagory("")
         }
+    }
+
+    const addCatagory = () => {
+        setAddCatagoryModalactive(true)
+        setModal(true)
+
+
+    }
+    const SubCatagoryAddbuttonFunction = () => {
+        addSubCatagory()
+        setModal(false);
+        setAddCatagoryModalactive(false);
+    }
+
+
+
+    const transactionAddButtonFunction = () => {
+        setModal(false)
+        setAddTrasactionModalActive(false)
+    }
+
+    const addTranactionBtn = () => {
+        setAddTrasactionModalActive(true);
+        setModal(true)
+
+    }
+
+    const cancelModal = () => {
+        setModal(false),
+            setAddCatagoryModalactive(false);
+        setAddTrasactionModalActive(false)
     }
 
     console.log(subCatagories)
@@ -30,33 +60,89 @@ const Income = () => {
                 Income
             </Text>
 
-            <ModalForUser />
-            <View style={styles.AddbtnContainer}>
-                <TouchableOpacity style={styles.addTrasactionBtn}>
-                    <Text style={styles.btnText}>
-                        Add trasaction
-                    </Text>
-                </TouchableOpacity>
-                <TouchableOpacity style={styles.addSubCatagorynBtn} onPress={() => setModal(true)}>
-                    <Text style={styles.btnText}>
-                        Add Sub Catagory
-                    </Text>
-                </TouchableOpacity>
+            {/* Add buttons */}
+            <View style={{ marginTop: 10, width: "40%" }}>
+                <Button
+                    onPress={() => addCatagory()}
+                    title="Add catagory"
+                    color="#262A53"
+                />
+            </View>
+            <View style={{ marginTop: 10, width: "40%", }}>
+                <Button
+                    title="Add transaction"
+                    color="#38A3A5"
+                    onPress={() => addTranactionBtn()}
+                />
             </View>
 
-            {/* <View style={addmoreCatagoryWrapper}>
-                <TextInput
-                    style={styles.newSubCatagory}
-                    placeholder="Add more sub catagory"
-                    onChangeText={text => setNewCatagory(text)}
-                    value={newCatagory}
+            {/* Sub catagories */}
+            {
+                subCatagories ? <View>
+                    <FlatList
+                        style={styles.catagoryList}
+                        horizontal={true}
+                        data={subCatagories}
+                        keyExtractor={(item, index) => index.toString()}
+                        renderItem={({ item }) =>
+                            <IncomeCatagories item={item} />
+                        }
+                    />
+                </View> : null
+            }
+
+
+            {/* Add sub catagory Modal */}
+            {addCatagoryModalactive ? <ModalForUser>
+                <Button
+                    onPress={() => cancelModal()}
+                    title="Cancel modal"
+                    color="red"
+                    accessibilityLabel="Learn more about this purple button"
                 />
-                <TouchableOpacity onPress={addSubCatagory}>
-                    <Text>
-                        Add sub catagory
-                    </Text>
-                </TouchableOpacity>
-            </View> */}
+                <Text style={{ textAlign: "center" }}>Add Sub catagory</Text>
+                <KeyboardAvoidingView>
+                    <View style={styles.addmoreCatagoryWrapper}>
+                        <TextInput
+                            style={styles.newSubCatagoryInput}
+                            placeholder="Add more sub catagory"
+                            placeholderTextColor="#787A91"
+                            onChangeText={text => setNewCatagory(text)}
+                            value={newCatagory}
+                        />
+                        <TouchableOpacity style={styles.subCatagoryAddBtn}
+                            onPress={() => SubCatagoryAddbuttonFunction()}>
+                            <Text style={styles.subCatagoryAddBtnText}>
+                                Save
+                            </Text>
+                        </TouchableOpacity>
+                    </View>
+                </KeyboardAvoidingView>
+            </ModalForUser> : null}
+
+            {/* Add transaction Modal */}
+            {addTrasactionModalActive ? <ModalForUser>
+                <Button
+                    onPress={() => cancelModal()}
+                    title="Cancel modal"
+                    color="red"
+                    accessibilityLabel="Learn more about this purple button"
+                />
+                <Text style={{ textAlign: "center" }}>Add new Transaction</Text>
+
+                <View style={styles.addTransactionContainer}>
+                    <TextInput
+                        style={styles.addTransactionInput}
+                        placeholder="Amount"
+                    />
+                    <TouchableOpacity onPress={() => transactionAddButtonFunction()} style={styles.addTransactionBtn}>
+                        <Text style={styles.addTransactionBtnText}>
+                            Save
+                        </Text>
+                    </TouchableOpacity>
+                </View>
+            </ModalForUser> : null}
+
         </View>
 
     )
@@ -68,7 +154,8 @@ const styles = StyleSheet.create({
     container: {
         flex: 1,
         marginTop: StatusBar.currentHeight,
-        backgroundColor: "white"
+        backgroundColor: "white",
+        paddingHorizontal: 10,
     },
     pageTitle: {
         textAlign: "center",
@@ -76,34 +163,62 @@ const styles = StyleSheet.create({
         fontWeight: "800"
 
     },
-    AddbtnContainer: {
-        position: "absolute",
+    catagoryList: {
         flexDirection: "row",
-        alignSelf: "center",
-        width: "90%",
-        justifyContent: "space-around",
-        bottom: 10,
+        borderRadius: 10,
+        paddingHorizontal: 10,
+        backgroundColor: "#FFE3E3",
+        height: 40,
     },
-    addTrasactionBtn: {
-        backgroundColor: "#DF2E2E",
-        height: 50,
+    addmoreCatagoryWrapper: {
+        alignItems: "center",
+        marginBottom: 10,
+    },
+    newSubCatagoryInput: {
+        marginTop: 10,
+        backgroundColor: "#EEEEEE",
+        width: "100%",
+        height: 40,
+        padding: 10,
+        fontSize: 15,
+        borderRadius: 10,
+    },
+    subCatagoryAddBtn: {
+        marginTop: 10,
+        backgroundColor: "#262A53",
+        height: 40,
         borderRadius: 5,
-        padding: 8,
+        width: "100%",
+        alignItems: "center",
         justifyContent: "center"
+    },
+    subCatagoryAddBtnText: {
+        color: "white",
+        fontSize: 15,
+    },
+    addTransactionContainer: {
 
     },
-    addSubCatagorynBtn: {
-        backgroundColor: "#3D087B",
+    addTransactionInput: {
+        marginTop: 10,
+        backgroundColor: "#EEEEEE",
+        width: "100%",
         height: 50,
-        borderRadius: 5,
-        padding: 8,
-        justifyContent: "center"
-
+        padding: 10,
+        fontSize: 15,
+        borderRadius: 10,
     },
-    btnText: {
-        color: "#fff",
-        fontSize: 18,
-
-
+    addTransactionBtn: {
+        marginTop: 10,
+        backgroundColor: "#262A53",
+        height: 40,
+        borderRadius: 5,
+        width: "100%",
+        alignItems: "center",
+        justifyContent: "center"
+    },
+    addTransactionBtnText: {
+        color: "white",
+        fontSize: 15,
     },
 })