+++ App/Screens/modal.js
@@ -1,11 +1,11 @@
 import React, { useContext, useEffect, useRef, useState } from 'react';
-import { Animated, Modal, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
+import { Animated, Modal, StyleSheet, View } from 'react-native';
 import { AppContext } from './../Context/AppContext';
 
 
 
 
-export default function ModalForUser() {
+export default function ModalForUser({children}) {
     const { modal, setModal } = useContext(AppContext)
     const [showModal, setShowModal] = useState(modal)
 
@@ -38,33 +38,8 @@ export default function ModalForUser() {
     return (
         <Modal transparent visible={showModal}>
             <View style={styles.modalBackGround}>
-
                 <Animated.View style={[styles.modalContianer, { transform: [{ scale: scaleValue }] }]}>
-                    <View style={{ alignItems: "center" }}>
-                        <View style={styles.headerBtn}>
-                            <TouchableOpacity onPress={() => setModal(false)}>
-                                {/* <Image
-                                    source={require('../assets/modalItems/x.png')}
-                                    style={{ height: 30, width: 30 }}
-                                /> */}
-                            </TouchableOpacity>
-                        </View>
-                    </View>
-                    <View style={{ alignItems: "center" }}>
-                        <TouchableOpacity onPress={() => setModal(false)}>
-                            {/* <Image
-                                source={require('../assets/modalItems/success.png')}
-                                style={{ height: 100, width: 100 }}
-                            /> */}
-                             <Text style={{ marginVertical: 10, fontSize: 15, textAlign: "center" }}>
-                           Cancel
-                        </Text>
-                        </TouchableOpacity>
-
-                        <Text style={{ marginVertical: 10, fontSize: 15, textAlign: "center" }}>
-                            Successully added your Transaction.
-                        </Text>
-                    </View>
+                   {children }
                 </Animated.View>
             </View>
         </Modal>
@@ -79,7 +54,7 @@ const styles = StyleSheet.create({
         alignItems: 'center',
     },
     modalContianer: {
-        width: '60%',
+        // width: '70%',
         backgroundColor: 'white',
         paddingHorizontal: 20,
         paddingVertical: 20,